requirements.txt
Click>=6.0
click_repl
pathlib
jsonschema
stackprinter
jinja2
pyyaml
jsondiff
pytest
